✨Feat: Implement inbound approval list REST API
